Remove Ajax block on displaymode=page
In a project of ours we got an issue where we used certain categorypages for just content ("displaymode = page"). The page broke because the `web200.navigation.ajax` block expects products to be present.

It's very weird that Magento does this, since you've included this block only on `view_type_layered`, but apparently a `displaymode_page` category is still seen as `type_layered`.